+++ app/src/main/kotlin/me/xizzhu/android/joshua/bookmarks/BookmarksActivity.kt
@@ -18,6 +18,10 @@ package me.xizzhu.android.joshua.bookmarks
 
 import android.os.Bundle
 import me.xizzhu.android.joshua.R
+import me.xizzhu.android.joshua.bookmarks.list.BookmarksListView
+import me.xizzhu.android.joshua.bookmarks.list.BookmarksPresenter
+import me.xizzhu.android.joshua.bookmarks.toolbar.BookmarksToolbar
+import me.xizzhu.android.joshua.bookmarks.toolbar.ToolbarPresenter
 import me.xizzhu.android.joshua.ui.LoadingSpinner
 import me.xizzhu.android.joshua.ui.LoadingSpinnerPresenter
 import me.xizzhu.android.joshua.ui.bindView
@@ -32,9 +36,13 @@ class BookmarksActivity : BaseSettingsActivity() {
     lateinit var loadingSpinnerPresenter: LoadingSpinnerPresenter
 
     @Inject
+    lateinit var toolbarPresenter: ToolbarPresenter
+
+    @Inject
     lateinit var bookmarksPresenter: BookmarksPresenter
 
     private val loadingSpinner: LoadingSpinner by bindView(R.id.loading_spinner)
+    private val toolbar: BookmarksToolbar by bindView(R.id.toolbar)
     private val bookmarksListView: BookmarksListView by bindView(R.id.bookmarks)
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -42,6 +50,7 @@ class BookmarksActivity : BaseSettingsActivity() {
 
         setContentView(R.layout.activity_bookmarks)
         bookmarksListView.setPresenter(bookmarksPresenter)
+        toolbar.setPresenter(toolbarPresenter)
 
         observeSettings(bookmarksInteractor)
     }
@@ -50,11 +59,13 @@ class BookmarksActivity : BaseSettingsActivity() {
         super.onStart()
 
         loadingSpinnerPresenter.attachView(loadingSpinner)
+        toolbarPresenter.attachView(toolbar)
         bookmarksPresenter.attachView(bookmarksListView)
     }
 
     override fun onStop() {
         loadingSpinnerPresenter.detachView()
+        toolbarPresenter.detachView()
         bookmarksPresenter.detachView()
 
         super.onStop()