+++ app/src/main/kotlin/me/xizzhu/android/joshua/core/repository/BookmarkRepository.kt
@@ -17,11 +17,18 @@
 package me.xizzhu.android.joshua.core.repository
 
 import me.xizzhu.android.joshua.core.Bookmark
+import me.xizzhu.android.joshua.core.Constants
 import me.xizzhu.android.joshua.core.VerseIndex
 import me.xizzhu.android.joshua.core.repository.local.LocalBookmarkStorage
 
 class BookmarkRepository(private val localBookmarkStorage: LocalBookmarkStorage) {
-    suspend fun read(): List<Bookmark> = localBookmarkStorage.read()
+    suspend fun readSortOrder(): Int = localBookmarkStorage.readSortOrder()
+
+    suspend fun saveSortOrder(@Constants.SortOrder sortOrder: Int) {
+        localBookmarkStorage.saveSortOrder(sortOrder)
+    }
+
+    suspend fun read(@Constants.SortOrder sortOrder: Int): List<Bookmark> = localBookmarkStorage.read(sortOrder)
 
     suspend fun read(bookIndex: Int, chapterIndex: Int): List<Bookmark> = localBookmarkStorage.read(bookIndex, chapterIndex)