+++ app/src/main/kotlin/me/xizzhu/android/joshua/core/repository/local/android/db/BookmarkDao.kt
@@ -22,6 +22,7 @@ import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
 import androidx.annotation.WorkerThread
 import me.xizzhu.android.joshua.core.Bookmark
+import me.xizzhu.android.joshua.core.Constants
 import me.xizzhu.android.joshua.core.VerseIndex
 
 class BookmarkDao(private val sqliteHelper: SQLiteOpenHelper) {
@@ -45,10 +46,15 @@ class BookmarkDao(private val sqliteHelper: SQLiteOpenHelper) {
 
     private val db by lazy { sqliteHelper.writableDatabase }
 
-    fun read(): List<Bookmark> {
+    fun read(@Constants.SortOrder sortOrder: Int): List<Bookmark> {
         var cursor: Cursor? = null
         try {
-            cursor = db.query(TABLE_BOOKMARK, null, null, null, null, null, "$COLUMN_TIMESTAMP DESC")
+            val orderBy = when (sortOrder) {
+                Constants.SORT_BY_DATE -> "$COLUMN_TIMESTAMP DESC"
+                Constants.SORT_BY_BOOK -> "$COLUMN_BOOK_INDEX ASC, $COLUMN_CHAPTER_INDEX ASC, $COLUMN_VERSE_INDEX ASC"
+                else -> throw IllegalArgumentException("Unsupported sort order - $sortOrder")
+            }
+            cursor = db.query(TABLE_BOOKMARK, null, null, null, null, null, orderBy)
             return with(cursor) {
                 val bookmarks = ArrayList<Bookmark>(count)
                 val bookIndex = getColumnIndex(COLUMN_BOOK_INDEX)