+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG
 ---------
 
 #### Next Release
+- Supported sorting bookmarks by book
 
 #### v0.4.1 (2019-05-23)
 - Added missing on click listener for simple reading mode